Scan the key matrix
This commit sets up the column and row gpios, and performs a simple scan of a fixed keymap. This scan is triggered by a timer running at one kilohertz, and the result is used to populate the keyboard hid report. A report is not sent when the state of the keyboard has not changed.

Some delay is required between pulling a column down and pulling the next one up. Without it, there are false positives from the next column. The actual amount of delay hasn't been tuned.

This commit message is written using the keyboard running this firmware.