Implement a basic layer mechanism
This commit makes a first pass at a layering mechanism. The key matrix is first scanned, and the scan results are processed in a following step. If any of the keys are layer modifiers, it changes the layer, and the last processed layer key is the one that takes priority. Only momentary layers are implemented, but the provision for more complex layering constructs like toggle, set base, until next press etc has been set in place.

It's likely that this mechanism will need to be improved when other key matrix entry types are added, like combos, mod taps, and tap dances.