Implement mod-tap/tap-hold
This is the initial mod tap implementation. It has proper home row mods and works on multiple layers, although there is a difference: The hold interrupt comes in on a key press rather than a press and then release as in QMK. This does kind of give a generally slower feel to the whole thing, but it's workable, and means that when I go to capitalize with s or k, I'll often see a "di" pop up instead of an "I".