Change how interruptions are handled in tap-holds
Having looked more closely at how QMK handle tap-holds, the default
behaviour (and the one recommended for home row mods) is that keys
pressed during a tap hold that has not yet reached the timeout are
ignored, but do not cancel the taphold. This has a much better feel, and
indeed now feels pretty much the same as the qmk firmware.