Use 32-bit matrix entries to encode key data
This has the advantage of allowing normal keycodes to encode the full range and combination of modifiers, by dedicating 16-bits to the keycode. It gives a lot of additional space to extra functions like tap-hold - since they can be applied to modified keys like `S(KC_4)`, home row mods on all layers is trivial. It also allows the hold action to not encode modifiers, but also a key. With this, it's now possible to do thing like escape when tapped, backtick/tilde when held without needing the full tapdance functionality.

The tradeoffs made in qmk are sensible for a project targeting not only a wide range of processors, but also very limited and constrained processors. But in this project, and more powerful micros, using a full 32-bits per keymap entry is worth it.