Implement double taps
As an alternative to implementing the full tapdance functionality, I've
built "double tap", which allows two paths for a key action depending on
whether it was a single or a double tap. It might be worth combining the
functionality of double tap and taphold together at some point, but it
wouldn't be encodable in a single 32-bit keymap entry.