Add mechanism for adjusting the taphold hold time on a per-key basis
This allows a key to have a signed offset from the base value (e.g.
200ms), adjusting tapholds on weaker fingers to be longer and stronger
fingers to be shorter, depending on preference. To support a wider range
of hold delays, 16 bits are now dedicated to a millisecond counter
instead of 8.

This commit also removes the O+P combo (caps lock), as it was misfiring
too often.