Issue key pressed events to downstream modules from keyboard.c
This commit is the first step in moving the code for taphold, combo, and
double tap to a more event based structure. The taphold system is
refactored to work based on 3 "events" (loosely speaking):

1. press
2. release
3. update

update is just a function called by keyboard.c every scan, and is
intended to progress the system according to time passed.

All of the current systems should be able to work under this regime, and
being simpler and more efficent in terms of the total number of times
the entire key matrix is looped over.