Add settle time whenever setting the gpio state of a column during matrix scan
In a release build, I found that the entire first column wouldn't work. This turns out to be that there isn't enough time to see the signal make it through to the row pins. I had seen a similar effect when first implementing scanning, but it was that there was a residual row signal when deasserting the column pin. I've added an inline function with a configurable delay (measured in iterations of a `nop` in a for loop). Though I haven't measured it, the delay in the code now is probably something on the order of ~2us.