Refactor the combo system to work with press, release, and update
The combo system is without a doubt the most tricky and subtle of the 3.
In order to get this to work, and feel correct, I needed to ensure that
no combos are triggered while tapholds are active, otherwise a roll
involving a taphold and combo would always issue the combo key first,
despite the taphold key actually happening first in time.
This does mean that no taphold key can be part of a combo, but I think
that makes sense anyway.