Add a split layer
The split layer is accessible on either half using the outermost thumb
switch. It currently contains a few nicities for backspace and delate,
as well as dedicated copy paste for times when the right hand is
dedicated to mouse

This keyboard also appears to need to a longer settling period to avoid
spurious keys when typing quickly.