Implement a keyboard reconfiguration protocol
Similar to via or vial. This commit adds a new pair of vendor class
bulk endpoints and a simple packet-based protocol to get keyboard
information, as well as setting new mappings temporarily, or committed
to flash.

Further commits will extend the system to include extensions to support
macros and combos, as well as expanding the python program used to
reconfigure the keyboard.