Do most of the work of scanning and processing keys outsde of the ISR
Previously all the update logic was running inside the context of the
timer/alarm ISR callback. Though this wasn't an issue due to the short
processing time required, the timer now only sets a flag and has the
logic run in the regular context. Additionally the main loop waits for
interrupt before checking if the time has elapsed, which should reduce
power consumption.

This commit also fixes an issue where "invalid" combos would stop the
processing of any combos that came later. Previously this was OK, but
with the kbconfig system allowing for reconfigurable combos in any of
the slots, this needed to change.